WIP
build-macos.sh - Some minor changes to get CustomMKL to build on the mac.
I should probably combine this with build-linux.sh.

headers.txt - Had to add dnn headers needed by the newer version of MKL
I have to use because nobody knows how to download the mac version of
the old version that CNTK is currently using

For the next two files I'm just lifting things out of
https://github.com/Microsoft/CNTK/pull/100

MakeFile - Making the file output so or dylib depending on platform

configure - Had to change bash ! header to deal with the fact that
Apple ships bash 3 and we need bash 4. This change in ! let's us install
bash 4 on the mac in a way that shouldn't mess with linux (I tested it
and it seemed fine). Needed to change 'stat' to use -c on the mac not -f
which isn't supported.