Fixing compatibility issues as always
Makefile - readlink -f doesn't work on the mac. So I
'brew install coreutils' and use greadlink which does support -f. But this
required me to put in a switch so we use the right one.

BlockMultiplierPlatform.h - The Mac doesn't support aligned_alloc so I
found a few folks who had written wrappers online and synthesized them
into my own little call around posix_memalign.

PerformanceProfiler.cpp - Mac doesn't support syscall(SYS_gettid) to get
the current thread. So I found this work around. It has the obvious
problem that the returned value from the ID hash is a double but so
is the return value from GetCurrentThreadId in Windows and that isn't
bothering anyone that we compile it to an unsigned int.

configure - Cleaned up the apple logic a bit and restored readlink logic.