More fixes
Makefile - gcc-4.8 on the mac doens' tlike -rdynamic which is silly since
it's an optional switch. But it is only relevant to ELF binaries which
aren't used on the mac.
   The perf files won't compile without TimerUtility which wasn't included
in the file list. This is odd because they appear to compile on Linux just
fine. So for now I'm adding this as an optional addition just for the mac.

ScriptableObjects.h - I ran into problems with this line twice. At one
point the c++ compiler was unhappy because the auto value didn't match
the dynamic_cast that was originally used here. Changing from auto to C *
for p fixed that. But then we ran into a second case where the compiler
didn't think it could do a dynamic cast. So eventually I went back to
auto but used a reinterpret_cast. We'll see if that blows up.