Building mess
Makefile -
* I accidentally put quotes around greadlink, just bad habit
* The shared switch for the mac is different than for everywhere else so
I had to put in a variable to handle that. Also the build file plays
games where it builds shared libraries with missing symbols depending on
the fact that those symbols will show up when those shared libraries are
combined with other shared libraries that have the missing symbols.
Windows and Linux are o.k. with this but the Mac isn't. It defaults to
returning errors if a symbol isn't resolvable when building a shared
library. To work around this I had to add in the undefined argument
with dynamic_lookup. We'll see how far that gets me.
* I am playing around with rpath since near as I can tell $ORIGIN is
replaced with @executable_path in Mac land but I'm not 100% sure.
* I removed places where I was manually adding cpp files to the compile
instructions. This was done to deal with unresolved symbol errors that
I now try to address via the -undefined switch described above.