Bunch o' fixes
Makefile -
* The -Wno flags are to turn off errors that make LLVM unhappy.
* The DARWIN_C_SOURCE is because without this we don't get pthread.h
functions we need.
* The CXXFLAGS change as well as removing CPPFLAGS and LDFLAGS is
because the Makefile was ignoring environment variables. This seems wrong
in general but is a real problem on the Mac because we need those variables
to point to our openMP supporting custom installed LLVM tool chain.
* The -O4 to -O3 change is because there isn't a -O4, it just turns into
-O3 but while GCC just silently does this, LLVM throws an error

Basics.h - LLVM really doesn't like it when we overload a function with
a static and non-static version (this is actually illegal in the C++
standard).

File.h, TensorShape.h - LLVM doesn't like it when braces aren't included

ssematrix.h - LLVM hates the two functions because they defined f as const
HANDLE but HANDLER is really just void * and that doesn't match the
FILE *f expected by the f functions. So I if def'd them out to see if
they are actually being used.

RecurrentNodes.h - LLVM doesn't like /* inside of /*

CorpusDescriptor.h - PRIu64 is the wrong size on mac so instead I use
zu which if I'm not confused (and I probably am) should handle size_t
everywhere.

Criterion.h - LLVM doesn't like the absence of the template keyword to
disambiguate the class call.

setup_mac_environment.sh - THIS DOESN'T WORK AS WRITTEN! It's just notes
I'm using.